refactor: map 관련 컴포넌트 위치 이관작업
- components 서브 디렉토리 추가 후 이동
