+++ app/(main)/page.tsx
@@ -1,4 +1,4 @@
-import TopSection from '@/features/map/top-section'
+import TopSection from '@/features/map/components/top-section'
 import KakaoMap from '@/shared/components/common/kakao-map'
 
 export default function MainPage() {
